+++ tango/integrations/torch/train.py
@@ -204,9 +204,9 @@ class TorchTrainStep(Step):
             is_distributed = True
             num_workers = len(devices)
 
-        if train_steps is not None and train_epochs is not None:
+        if (train_steps is not None) == (train_epochs is not None):
             raise ConfigurationError(
-                "You cannot specify train_steps and train_epochs at the same time."
+                "One of 'train_steps' or 'train_epochs' needs to be specified, but not both."
             )
 
         config = TrainConfig(
@@ -336,13 +336,19 @@ def _train(
     _check_dataset(train_dataset, config.train_split)
     train_dataloader: DataLoader = train_dataloader.construct(dataset=train_dataset)
 
-    assert config.train_steps is None or config.train_epochs is None
-    if config.train_steps is None:
-        try:
-            config.train_steps = len(train_dataloader) * (config.train_epochs or 1)
-        except TypeError:
+    steps_per_epoch: Optional[int] = None
+    try:
+        steps_per_epoch = len(train_dataloader)
+    except TypeError:
+        if config.train_steps is None:
             raise ConfigurationError("You must set 'train_steps' for streaming/iterable datasets")
+
+    if config.train_steps is None:
+        assert steps_per_epoch is not None and config.train_epochs is not None
+        config.train_steps = steps_per_epoch * (config.train_epochs or 1)
+
     assert config.train_steps is not None  # for mypy
+
     if validation_dataloader is not None:
         if config.validation_steps is None:
             try:
@@ -519,7 +525,7 @@ def _train(
             total=start_step - 1,
             disable=not config.is_local_main_process,
         ) as batch_iter:
-            for step, batch in batch_iter:
+            for step, (_, batch) in batch_iter:
                 del batch
                 if step >= start_step - 1:
                     break
@@ -538,9 +544,16 @@ def _train(
         disable=not config.is_local_main_process,
     )
     try:
-        for step, batch in train_batch_iterator:
+        for step, (epoch, batch) in train_batch_iterator:
+            # Pre-epoch callback.
+            if steps_per_epoch is not None and step % steps_per_epoch == 0:
+                for callback in callbacks:
+                    callback.pre_epoch(epoch)
+
+            # Pre-batch callback.
             for callback in callbacks:
                 callback.pre_batch(step, batch)
+
             optimizer.zero_grad()
             batch_loss = 0.0
             for micro_batch in batch:
@@ -567,6 +580,7 @@ def _train(
                 del outputs
                 del micro_batch_loss
 
+            # Post-batch callback.
             for callback in callbacks:
                 callback.post_batch(step, batch_loss)
 
@@ -686,6 +700,7 @@ def _train(
                 elif not config.minimize_val_metric and val_metric >= best_val_metric:
                     best_val_metric = val_metric
 
+                # Post validation callback.
                 for callback in callbacks:
                     callback.post_val_loop(step, val_metric, best_val_metric)
 
@@ -701,6 +716,11 @@ def _train(
             # Checkpoint.
             if config.should_checkpoint_this_step(step):
                 save_state(step)
+
+            # Post-epoch callback.
+            if steps_per_epoch is not None and (step + 1) % steps_per_epoch == 0:
+                for callback in callbacks:
+                    callback.post_epoch(epoch)
     except StopEarly:
         if config.is_local_main_process:
             print("Stopping early!")
@@ -725,7 +745,7 @@ def _cycle_through_epochs(dataloader: DataLoader, is_distributed: bool):
         if is_distributed and isinstance(dataloader.sampler, DistributedSampler):
             dataloader.sampler.set_epoch(epoch)
         for batch in dataloader:
-            yield batch
+            yield (epoch, batch)
         epoch += 1